\name{bibliography-class}
\alias{as.bibliography}
\alias{"[.bibliography"}
\alias{print.bibliography}
\alias{summary.bibliography}
\title{Methods for class 'bibliography'}
\description{This is a small number of standard methods for interacting with class 'bibliography'. More may be added later}
\section{Methods}{
Methods defined with class "bibliography" in the signature:
	\describe{
		\item{[}{select subset of entries}
		\item{as.bibliography}{convert data.frame to bibliography}
		\item{print}{print function}
		\item{summary}{summary function}
	}
}