\name{bibliography-methods}
\title{Methods for class 'bibliography'}
\alias{as.bibliography}
\alias{as.data.frame.bibliography}
\alias{print.bibliography}
\alias{summary.bibliography}
\alias{[.bibliography}
\alias{bibliography-methods}
\description{This is a small number of standard methods for interacting with class 'bibliography'. More may be added later}
\usage{
	as.bibliography(x, ...)
	as.data.frame(x, ...)
	\method{[}{bibliography}(x, n)
	\method{print}{bibliography}(x, n, ...)
	\method{summary}{bibliography}(object, ...)
}
\arguments{
  \item{x}{An object of class 'bibliography'}
  \item{object}{An object of class 'bibliography'}
  \item{n}{Number of items to select/print}
  \item{...}{Any further information}
}
