% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dtm.R
\name{make_dtm}
\alias{make_dtm}
\title{Construct a document-term matrix (DTM)}
\usage{
make_dtm(x, stop_words, min_freq = 0.01, max_freq = 0.85, verbose = FALSE)
}
\arguments{
\item{x}{a vector or \code{data.frame} containing text}

\item{stop_words}{optional vector of strings, listing terms to be removed
from the DTM prior to analysis. Defaults to \code{revwords()}.}

\item{min_freq}{minimum proportion of entries that a term must be found in
to be retained in the analysis. Defaults to 0.01.}

\item{max_freq}{maximum proportion of entries that a term must be found in
to be retained in the analysis. Defaults to 0.85.}

\item{verbose}{logical: should details be printed to the screen? Defaults to FALSE}
}
\value{
A list as described in \code{link{stm::prepDocuments}}
}
\description{
Takes bibliographic data and converts it to a DTM for passing to topic
models.
}
\details{
Formerly a unique function, but now a wrapper to code{quanteda} functions. Maintained for backward-compatability.
}
\examples{

# import some data
file_location <- system.file(
  "extdata",
  "avian_ecology_bibliography.ris",
  package = "revtools")
x <- read_bibliography(file_location)

# construct a document-term matrix
# note: this can take a long time to run for large datasets
x_dtm <- make_dtm(x$title)

# to return a matrix where length(output) == nrow(x)
x_dtm <- make_dtm(x$title, retain_empty_rows = TRUE)
x_dtm <- as.matrix(x_dtm) # to convert to class 'matrix'
dim(x_dtm) # 20 articles, 10 words

}
\seealso{
\code{\link{run_topic_model}}, \code{\link{screen_topics}}
}
