% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preload_screen_abstracts.R
\name{preload_screen_abstracts}
\alias{preload_screen_abstracts}
\title{Preload an app for screening article abstracts}
\usage{
preload_screen_abstracts(data, file, write, app_control)
}
\arguments{
\item{data}{A \code{data.frame} containing bibliographic information.}

\item{file}{Name of the output \code{rds} file. Defaults to \code{'screen_abstracts.rds'}}

\item{write}{Logical; should the specified \code{file} be written? Defaults to TRUE; if FALSE the
resulting app is returned to the workspace rather than the specified filename.}

\item{app_control}{A list containing additional information to customize the UI.
See details for what options can be set.}
}
\value{
Returns file or object (as specified by the \code{format} argument)
containing the information needed to build an app with the required
specifications. Specifically, it returns an S3 object of class
\code{screen_abstracts_preloaded} containing three slots: \code{data}, which
stores the underlying dataset including any user decisions, notes and timing
information; \code{file} which stores the output file name; and
\code{app_control}.
}
\description{
This function takes some data and generates a self-contained app, stored in
a \code{.rds} file, that can be forwarded to collaborators for screening. A
series of options can be pre-set to customize user experience.
}
\details{
List entries that can be specified within \code{app_control} are as follows:
\itemize{
\item\code{show_identifying_info} - Logical: should author names,
publication year and journal title be displayed? Defaults to FALSE.
\item\code{time_responses} - Logical: should the app record timing
information? If TRUE (the default) record the time spent on each page, and
the time at which the page was loaded.
\item\code{keyword_highlighting} - Logical: should the app highlight
user-specified keywords? Defaults to FALSE.
\item\code{keywords} - Optional set of keywords to be highlighted in
article titles and abstracts.  \item\code{highlight_color} - Optional color
for keyword highlighting. Defaults to "red".
\item\code{rank_by} - Specify how entries should be ordered. Should be one
of \code{"initial"} (the default) to maintain the order of \code{x},
\code{"random"} to randomize order via \code{rnorm}, \code{"alphabetical"}
for alphabetical by article title, or \code{"relevance"} for ranking by
number of matched keywords.
\item\code{save_csv} - Logical: should the app save a \code{csv} file as
well as a \code{.rds} file? Defaults to FALSE.
}
}
\examples{

file_location <- system.file(
  "extdata",
  "avian_ecology_bibliography.ris",
  package = "revtools")
x <- read_bibliography(file_location)

\dontrun{
preload_screen_abstracts(x,
  file = "screening_app.rds",
  app_control = list(
    keyword_highlighting = TRUE,
    highlight_color = "red",
    keywords = c("management", "frog", "bird", "fire"),
    save_csv = FALSE
  )
)
# then run the resulting app from file
screen_abstracts("screening_app.rds")
}

}
