\docType{package}
\name{revtools-package}
\alias{revtools-package}
\alias{revtools}
\title{revtools}
\description{Tools to support evidence synthesis}
\section{Functions}{
	revtools contains the following functions:
	\itemize{
		\item Import & export:
		\item \code{\link{read_bibliography}} Import bibliographic data
		\item \code{\link{write_bibliography}} Export bibliographic data
	}
}