% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_topic_model.R
\name{run_topic_model}
\alias{run_topic_model}
\title{Calculate a topic model}
\usage{
run_topic_model(dtm, n_topics = 5, iterations = 2000, ...)
}
\arguments{
\item{dtm}{a dfm as produced by \code{quanteda::dfm()} or equivalently
by \code{revtools::make_dtm()}.}

\item{n_topics}{Number of topics to calculate}

\item{iterations}{The maximum number of iterations}
}
\value{
A topic model with the specified parameters.
}
\description{
Calculate a topic model using the \code{stm} package.
}
\note{
This is synonymous with \code{stm::stm()}, but is kept for backwards-compatibility
with earlier versions of revtools.
}
\examples{

# import data
file_location <- system.file(
  "extdata",
  "avian_ecology_bibliography.ris",
  package = "revtools"
)
x <- read_bibliography(file_location)

# run a topic model based on these data
# note: the following lines can take a very long time to run, especially for large datasets!
x_dtm <- make_dtm(x$title)
\dontrun{x_lda <- run_topic_model(x_dtm, 5, 5000)}

}
\seealso{
\code{\link{make_dtm}} for constructing data to pass to this
function; \code{\link{screen_topics}} for interactive visualisation of topic
model results.
}
