% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_topics_progress.R
\name{summary.screen_topics_progress}
\alias{summary.screen_topics_progress}
\alias{screen_topics_progress-methods}
\title{Methods for class 'screen_topics_progress'}
\usage{
\method{summary}{screen_topics_progress}(object, ...)
}
\arguments{
\item{object}{An object of class 'screen_topics_progress'}

\item{...}{Any further information}
}
\value{
Prints useful information to the workspace.
}
\description{
Tools to display useful information on class \code{screen_topics_progress}.
}
\note{
Class \code{screen_topics_progress} is a format for exporting large
quantities of data during reviews. It is typically stored within a .rds file
in the working directory. When re-imported to R using \code{readRDS}, this
file will contain an object of class \code{screen_topics_progress}.
}
