\name{bibliography-class}
\alias{as.bibliography}
\alias{'[.bibliography'}
\alias{print.bibliography}
\alias{summary.bibliography}
\title{Methods for class 'bibliography'}
\description{This is a small number of standard methods for interacting with class 'bibliography'. More may be added later}
\usage{
	as.bibliography(x)
	\method{[}{bibliography}(x, \dots)
	\method{print}{bibliography}(x, n)
	\method{summary}{bibliography}
}
\arguments{
  \item{x}{An object of class 'bibliography'}
  \item{n}{Number of items to print}
  \item{...}{Any further information}
}
