% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{read_bibliography}
\alias{read_bibliography}
\title{Import bibliographic data}
\usage{
read_bibliography(
  filename,
  return_df = TRUE,
  tag_naming = "best_guess",
  verbose = FALSE
)
}
\arguments{
\item{filename}{A vector or list containing paths to one or more
bibliographic files.}

\item{return_df}{Logical; should the object returned be a data.frame?
Defaults to TRUE. If FALSE, returns an object of class \code{bibliography}.}

\item{tag_naming}{Specify the type of ris tags used.}

\item{verbose}{Logical; should progress information be printed?}
}
\value{
Returns an object of class \code{data.frame} if \code{return_df} is
\code{TRUE}; otherwise an object of class \code{bibliography}.
}
\description{
Import standard formats from academic search engines and referencing
software.
}
\details{
This function aims to import bibliographic data from a range of formats in a
consistent manner. \strong{This function has been deprecated and will be removed
from future versions of \code{revtools}. Please use \code{synthesisr::read_refs()}
instead}.
}
\examples{

file_location <- system.file(
  "extdata",
  "avian_ecology_bibliography.ris",
  package = "revtools")
x <- read_bibliography(file_location)
class(x) # = data.frame
x <- read_bibliography(file_location, return_df = FALSE)
class(x) # = bibliography
summary(x)

}
\seealso{
\code{\link{bibliography-class}}, \code{\link{tag_lookup}}
}
