% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{write_bibliography}
\alias{write_bibliography}
\title{Export imported bibliographic data as .bib or .ris formats}
\usage{
write_bibliography(x, filename, format = "ris", tag_naming = "synthesisr")
}
\arguments{
\item{x}{A \code{data.frame} or object of class \code{bibliography}}

\item{filename}{Name of the exported file. Should ideally match 'format',
but this is not enforced}

\item{format}{Format of the exported file. Should be either "ris" (default)
or "bib"}

\item{tag_naming}{Type of tags to use in writing the output file}
}
\value{
exports results as a .ris or .bib file.
}
\description{
Basic function to export bibliographic information for use in other
programs. \strong{This function has been deprecated and will be removed
from future versions of \code{revtools}. Please use \code{synthesisr::write_refs()} 
instead}.
}
\examples{

file_location <- system.file(
  "extdata",
  "avian_ecology_bibliography.ris",
  package = "revtools")
x <- read_bibliography(file_location, return_df = FALSE)

# export a subset of entries as a new file
write_bibliography(x[1:5],
  filename = paste0(tempdir(), "/x_out.ris"),
  format = "ris")

}
